footnotes related code is factored out
